# Backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BeeBa API
  version: 0.1.0
  description: Versioned API for the BeeBa Basis / VR content catalog.
servers:
  - url: http://localhost:8080
    description: Local backend
  - url: https://beeba.org
    description: Production domain
paths:
...
  /api/v1/users/{username}:
    get:
      summary: Get a public user profile
      operationId: getPublicProfile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 32
      responses:
        "200":
          description: Public user profile with published-content counters.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/PublicProfile"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/search:
    get:
      summary: Search published public content with Meilisearch
      operationId: searchContent
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
        - name: category
          in: query
          schema:
            type: string
            enum:
              - worlds
              - avatars
              - props
              - prefabs
        - name: tags
          in: query
          description: Comma-separated tag slugs. Results match any provided tag.
          schema:
            type: string
            maxLength: 500
        - name: include_nsfw
          in: query
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - newest
              - likes
              - downloads
            default: newest
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 24
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results hydrated from PostgreSQL after Meilisearch ranking.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicContentItem"
                  pagination:
                    $ref: "#/components/schemas/SearchPagination"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/content/{content_id}/report:
    post:
      summary: Report published public content
      operationId: reportContent
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportContentRequest"
      responses:
        "201":
          description: Open content report.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/ContentReport"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: beeba_access_token
      description: Browser session cookie. Unsafe methods authenticated by cookie also require `X-CSRF-Token` matching the `beeba_csrf_token` cookie.
  responses:
    Error:
      description: Consistent API error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    PublicContentItem:
      type: object
      required:
        - id
        - slug
        - title
        - description
        - status
        - visibility
        - nsfw
        - category
        - author
        - tags
        - likes_count
        - downloads_count
        - comments_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - published
        visibility:
          type: string
          enum:
            - public
        nsfw:
          type: boolean
        unlock_password:
          type: string
          description: Public Basis unlock password for published public catalog items.
        category:
          type: object
          required:
            - slug
            - name
          properties:
            slug:
              type: string
            name:
              type: string
        author:
          type: object
          required:
            - id
            - username
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            display_name:
              type:
                - string
                - "null"
            avatar_image_id:
              type:
                - string
                - "null"
              format: uuid
        tags:
          type: array
          items:
            type: object
            required:
              - slug
              - name
            properties:
              slug:
                type: string
              name:
                type: string
        preview_image_id:
          type:
            - string
            - "null"
          format: uuid
        likes_count:
          type: integer
          minimum: 0
        downloads_count:
          type: integer
          minimum: 0
        comments_count:
          type: integer
          minimum: 0
        published_at:
          type:
            - string
            - "null"
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PublicProfile:
      type: object
      required:
        - id
        - username
        - published_count
        - likes_count
        - downloads_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type:
            - string
            - "null"
        avatar_image_id:
          type:
            - string
            - "null"
          format: uuid
        published_count:
          type: integer
          minimum: 0
        likes_count:
          type: integer
          minimum: 0
        downloads_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ReportContentRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 120
        details:
          type: string
          maxLength: 2000
    ContentReport:
      type: object
      required:
        - id
        - content_id
        - user_id
        - reason
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        content_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        reason:
          type: string
        details:
          type:
            - string
            - "null"
        status:
          type: string
          enum:
            - open
            - in_review
            - resolved
            - rejected
        created_at:
          type: string
          format: date-time
...
    SearchPagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        next_cursor:
          type:
            - string
            - "null"
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
          maximum: 50
...